mka: incomplete for unified version
git-svn-id: https://localhost:8443/svn/expectance@261 aa805ad0-02f7-422c-84f2-0395ce76bb97

+++ trunk/src/engine/src/scene/renderobject.cpp
@@ -23,7 +23,7 @@ bool RenderObject::RenderShadow()
 	return true;
 }
 
-const byte RenderObject::getRenderingOrder()
+const short RenderObject::getRenderingOrder()
 {
 	return m_renderingOrder; 
 }
